move service creation back into replicaset controllger file
